Remove existing API key impl and replace it with new API key impl
