+++ integration/test-integration/src/test/resources/testng-cc-with-apim.xml
@@ -59,10 +59,6 @@
            <class name="org.wso2.choreo.connect.tests.testcases.withapim.ExistingApiTestCase"/>
            <class name="org.wso2.choreo.connect.tests.testcases.withapim.BackendSecurityTestCase"/>
            <class name="org.wso2.choreo.connect.tests.testcases.withapim.jwtValidator.JwtAndScopeTestCase"/>
-           <class name="org.wso2.choreo.connect.tests.testcases.withapim.apikey.APIKeyTestCase"/>
-<!--           <class name="org.wso2.choreo.connect.tests.testcases.withapim.apikey.APIKeyBlockedAPITestCase"/>-->
-           <class name="org.wso2.choreo.connect.tests.testcases.withapim.apikey.APIKeyAppLevelThrottleTestCase"/>
-           <class name="org.wso2.choreo.connect.tests.testcases.withapim.apikey.APIKeySubLevelThrottleTestCase"/>
            <class name="org.wso2.choreo.connect.tests.testcases.withapim.RetryAndTimeoutTestCase"/>
            <class name="org.wso2.choreo.connect.tests.testcases.withapim.EnvoyHttpFilterTestCase"/>
        </classes>
